backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic-settings>=2.7.0
pydantic[email]>=2.0.0
redis[asyncio]>=5.2.0
python-jose[cryptography]>=3.3.3
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
python-multipart>=0.0.20
httpx>=0.28.0
meilisearch>=0.31.0
boto3>=1.35.0
pypdf>=5.1.0
